Adding Amharic Language
Amharic Language is Spoken by Ethiopians. 
Ethiopia is the only country in Africa with its own distinct alphabet. 
It is the Major working language of Ethiopia ( >110M people )
There are millions of Ethiopians is USA and Europe
This language has a letter for each and every sound 
Here is the list of about 98% of the full alphabet letters. 

i tried it but  langdetect is not detecting it. i hope you'll find a way to add it